third_party/WebKit/LayoutTests/html5lib/generated/run-tests11-data-expected.txt

CONSOLE ERROR: line 1: Error: <svg> attribute viewBox: Unexpected end of attribute. Expected number, "".
CONSOLE ERROR: line 1: Error: <svg> attribute viewBox: Unexpected end of attribute. Expected number, "".
CONSOLE ERROR: line 1: Error: <svg> attribute viewBox: Unexpected end of attribute. Expected number, "".
../resources/tests11.dat: PASS
